+++ src/Symbology.php
@@ -24,6 +24,7 @@ enum Symbology: string
     case Code39 = 'code39';
     case Code39Extended = 'code39ext';
     case Code93 = 'code93';
+    case Codabar = 'codabar';
     case Ean13 = 'ean13';
     case Ean8 = 'ean8';
     case UpcA = 'upc-a';